requirements.txt
streamlit
Pillow
requests
matplotlib
plotly
numpy
scikit-learn
